moved spec folder to the root
